start all section functionality

+++ database/migrations/2022_05_20_015734_create_user_task_statuses_table.php
@@ -13,16 +13,18 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('user_task_statuses', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('task_id');
-            $table->unsignedBigInteger('user_id');
-            $table->boolean('status')->default(false);
-            $table->timestamps();
+        if (!Schema::hasTable('user_task_statuses')) {
+            Schema::create('user_task_statuses', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('task_id');
+                $table->unsignedBigInteger('user_id');
+                $table->boolean('status')->default(false);
+                $table->timestamps();
 
-            $table->foreign('task_id')->references('id')->on('tasks')->onDelete('cascade');
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-        });
+                $table->foreign('task_id')->references('id')->on('tasks')->onDelete('cascade');
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+            });
+        }
     }
 
     /**
